Organize TS code into separte files for future testing and avoid conflicts
