+++ src/ts/call-sass.ts
@@ -1,0 +1,1 @@
+import "../sass/main.scss";
